chore: rename test example
