+++ apps/test-examples/App.js
@@ -12,7 +12,7 @@ import Test1790 from './src/Test1790';
 import Test1813 from './src/Test1813';
 import Test1845 from './src/Test1845';
 import Test1986 from './src/Test1986';
-import Test2054 from './src/Test2054';
+import Test2397 from './src/Test2397';
 import Test2071 from './src/Test2071';
 import Test2080 from './src/Test2080';
 import Test2086 from './src/Test2086';
